backend/requirements.txt
requests>=2.32.0
pandas>=2.2.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dateutil>=2.9.0
